backend/pyproject.toml
[project]
name = "create-hima-app-backend"
version = "0.1.0"
description = "Create Hima App Backend"
authors = [
    { name = "vi.channnnn", email = "violet@himaa.me" },
]
dependencies = [
    "alembic==1.13.1",
    "asyncpg==0.29.0",
    "fastapi==0.109.1",
    "fastapi-pagination==0.12.5",
    "passlib==1.7.4",
    "python-dotenv==1.0.0",
    "python-jose==3.3.0",
    "python-multipart",
    "pytest==7.4.0",
    "pytest-asyncio==0.21.0",
    "pydantic==2.6.0",
    "pydantic[email]==2.6.0",
    "PyJWT==2.8.0",
    "uvicorn==0.27.0",
    "sqlalchemy==2.0.25",
    "httpx==0.26.0",
    "requests==2.31.0",
    "types-requests==2.31.0.1",
    "isort==5.12.0",
    "pylint==2.13.9",
    "mypy==1.8.0",
    "ruff==0.2.0",
    "pylint-flask-sqlalchemy",
    "sqlalchemy[mypy]",
    "types-python-jose",
    "types-passlib",
    "faker==15.3.4",
    "coverage==7.4.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true